% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuyina.map-package.R
\docType{package}
\name{nuyina.map-package}
\alias{nuyina.map}
\alias{nuyina.map-package}
\title{nuyina.map: Just a Map Around Nuyina}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/nuyina.map}
  \item Report bugs at \url{https://github.com/mdsumner/nuyina.map/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/0000-0002-2471-7511}{ORCID})

}
\keyword{internal}
