% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuyina_latest.R
\name{nuyina_latest}
\alias{nuyina_latest}
\title{Latest Nuyina underway}
\usage{
nuyina_latest(n = 3 * 24 * 3600)
}
\arguments{
\item{n}{seconds since latest}
}
\value{
data frame
}
\description{
Latest Nuyina underway
}
\examples{
nuyina_latest()
}
